上传医生照片

+++ chsis/src/main/java/top/chsis/service/impl/DepartmentServiceImpl.java
@@ -1,12 +1,19 @@
 package top.chsis.service.impl;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
+
 import top.chsis.dao.DepartmentMapper;
+import top.chsis.dao.DoctorMapper;
 import top.chsis.model.Department;
+import top.chsis.model.Doctor;
+import top.chsis.model.HospitalManager;
 import top.chsis.service.IDepartmentService;
 
 @Service("departmentService")
@@ -15,8 +22,18 @@ public class DepartmentServiceImpl implements IDepartmentService {
 	@Autowired
 	private DepartmentMapper departmentMapper;
 	
+	@Autowired
+	private DoctorMapper doctorMapper;
+	
 	public int deleteByPrimaryKey(String uuid) {
-		return departmentMapper.deleteByPrimaryKey(uuid);
+		List<Doctor> doctors = doctorMapper.selectDoctorsByDepartmentUUID(uuid);
+		
+		if(doctors !=null && doctors.size() > 0){
+			return 0;
+		}else{
+			return departmentMapper.deleteByPrimaryKey(uuid);
+		}
+		
 	}
 
 	public int insert(Department record) {
@@ -45,6 +62,18 @@ public class DepartmentServiceImpl implements IDepartmentService {
 
 	public List<Department> selectAll() {
 		return departmentMapper.selectAll();
+	}
+
+	public PageInfo<Department> selectByConditionAndPaging(Department department, int page, int size) {
+		PageHelper.startPage(page, size);
+		List<Department> list = new ArrayList<Department>();
+		list = departmentMapper.selectByCondition(department);
+		PageInfo<Department> pageInfo = new PageInfo<Department>(list);
+		return pageInfo;
+	}
+
+	public Department selectByNumber(String number) {
+		return departmentMapper.selectByNumber(number);
 	}
 
 }
