+++ chsis/src/main/java/top/chsis/service/impl/HospitalServiceImpl.java
@@ -5,9 +5,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import top.chsis.dao.DepartmentMapper;
 import top.chsis.dao.HospitalManagerMapper;
 import top.chsis.dao.HospitalMapper;
 import top.chsis.dao.ManagerMapper;
+import top.chsis.model.Department;
 import top.chsis.model.Hospital;
 import top.chsis.model.HospitalManager;
 import top.chsis.service.IHospitalService;
@@ -24,12 +26,21 @@ public class HospitalServiceImpl implements IHospitalService{
 	@Autowired
 	private ManagerMapper managerMapper;
 	
+	@Autowired
+	private DepartmentMapper departmentMapper;
+	
 	public int deleteByPrimaryKey(String uuid) {
-		List<HospitalManager> hospitalManagers = hospitalManagerMapper.selectByHospitalUuid(uuid);
-		for(HospitalManager hm : hospitalManagers) {
-			managerMapper.deleteByPrimaryKey(hm.getManager().getUuid());
+		List<Department> departments = departmentMapper.selectDepartmentsByHospitalUUID(uuid);
+		
+		if(departments != null && departments.size() > 0){
+			return 0;
+		}else{
+			List<HospitalManager> hospitalManagers = hospitalManagerMapper.selectByHospitalUuid(uuid);
+			for(HospitalManager hm : hospitalManagers) {
+				managerMapper.deleteByPrimaryKey(hm.getManager().getUuid());
+			}
+			return hospitalMapper.deleteByPrimaryKey(uuid);
 		}
-		return hospitalMapper.deleteByPrimaryKey(uuid);
 	}
 
 	public int insert(Hospital record) {
